Put final period utility in separate module
